Fix: token index data type
I wanted to save some memory with `uint8`, but `cross_entropy_loss`
expects target indices to be `long`.

+++ src/protornn/data.py
@@ -25,7 +25,7 @@ class ProteinSequenceDataset(Dataset):
 
     def __getitem__(self, idx) -> tuple[torch.Tensor, torch.Tensor]:
         sequence = self.sequences[idx]
-        dt = torch.uint8
+        dt = torch.long
 
         # Tokenize with BOS and EOS tokens
         input = self.tokenizer.encode(
